// src/types/enums.ts
export enum Role {
  ADMIN = 'ADMIN',
  DEPT_HEAD = 'DEPT_HEAD',
  DOCTOR = 'DOCTOR',
  NURSE = 'NURSE',
  RECEPTIONIST = 'RECEPTIONIST',
  PHARMACIST = 'PHARMACIST',
  LAB_TECHNICIAN = 'LAB_TECHNICIAN',
  PATIENT = 'PATIENT',
  ACCOUNTANT = 'ACCOUNTANT',
  PROGRAM_MANAGER = 'PROGRAM_MANAGER',
  CLINICIAN = 'CLINICIAN',
}
// ...
export enum Department {
  CARDIOLOGY = 'CARDIOLOGY',
  NEUROLOGY = 'NEUROLOGY',
  PEDIATRICS = 'PEDIATRICS',
  ORTHOPEDICS = 'ORTHOPEDICS',
  GYNECOLOGY = 'GYNECOLOGY',
  GENERAL_MEDICINE = 'GENERAL_MEDICINE',
  SURGERY = 'SURGERY',
  EMERGENCY = 'EMERGENCY',
  ONCOLOGY = 'ONCOLOGY',
  DERMATOLOGY = 'DERMATOLOGY',
  OPHTHALMOLOGY = 'OPHTHALMOLOGY',
  ENT = 'ENT',
  PSYCHIATRY = 'PSYCHIATRY',
  RADIOLOGY = 'RADIOLOGY',
  PATHOLOGY = 'PATHOLOGY',
  PHARMACY = 'PHARMACY',
  LABORATORY = 'LABORATORY',
  PHYSIOTHERAPY = 'PHYSIOTHERAPY',
}

export enum Shift {
  MORNING = 'MORNING',
  AFTERNOON = 'AFTERNOON',
  NIGHT = 'NIGHT',
  ROTATING = 'ROTATING',
}